chainlit_template/pyproject.toml
[project]
name = "chainlit-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mubashar-Bashir", email = "mubasharbashir003@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents[litellm]>=0.0.13",
]

[project.scripts]
chainlit-template = "chainlit_template:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
